+++ test/index.ts
@@ -10,5 +10,6 @@ import './theme/apply';
 import './storage/ThemeOption';
 import './pages/options/hooks/useTheme';
 import './theme/mirror';
+import './locales/translations';
 import './pages/notifications/components/Alert/Alert';
 import './pages/options/components/Options/Modal/Modal';